Move util classes
- Create package for util classes and move them there
