+++ src/main/java/org/phoebus/channelfinder/common/CFResourceDescriptors.java
@@ -1,4 +1,4 @@
-package org.phoebus.channelfinder;
+package org.phoebus.channelfinder.common;
 
 public class CFResourceDescriptors {
