+++ src/main/java/org/phoebus/channelfinder/respository/ChannelRepository.java
@@ -49,8 +49,8 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 import javax.annotation.PreDestroy;
-import org.phoebus.channelfinder.CFResourceDescriptors;
-import org.phoebus.channelfinder.TextUtil;
+import org.phoebus.channelfinder.common.CFResourceDescriptors;
+import org.phoebus.channelfinder.common.TextUtil;
 import org.phoebus.channelfinder.configuration.ElasticConfig;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
