+++ src/test/java/org/phoebus/channelfinder/docker/ChannelFinderScrollIT.java
@@ -24,6 +24,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 import java.net.HttpURLConnection;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
+import org.phoebus.channelfinder.common.CFResourceDescriptors;
 import org.phoebus.channelfinder.entity.Scroll;
 import org.phoebus.channelfinder.rest.controller.ChannelScroll;
 import org.testcontainers.containers.ComposeContainer;
@@ -32,7 +33,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 /**
  * Integration tests for ChannelFinder and Elasticsearch with focus on usage of {@link
- * org.phoebus.channelfinder.CFResourceDescriptors#SCROLL_RESOURCE_URI}.
+ * CFResourceDescriptors#SCROLL_RESOURCE_URI}.
  *
  * @author Lars Johansson
  * @see ChannelScroll
@@ -86,7 +87,7 @@ class ChannelFinderScrollIT {
     }
   }
 
-  /** Test {@link org.phoebus.channelfinder.CFResourceDescriptors#SCROLL_RESOURCE_URI}. */
+  /** Test {@link CFResourceDescriptors#SCROLL_RESOURCE_URI}. */
   @Test
   void handleScrollQueryChannels() {
     // what
@@ -102,7 +103,7 @@ class ChannelFinderScrollIT {
     }
   }
 
-  /** Test {@link org.phoebus.channelfinder.CFResourceDescriptors#SCROLL_RESOURCE_URI}. */
+  /** Test {@link CFResourceDescriptors#SCROLL_RESOURCE_URI}. */
   @Test
   void handleScrollContinueChannelsQuery() {
     // what
@@ -117,7 +118,7 @@ class ChannelFinderScrollIT {
   }
 
   /**
-   * Test {@link org.phoebus.channelfinder.CFResourceDescriptors#SCROLL_RESOURCE_URI}.
+   * Test {@link CFResourceDescriptors#SCROLL_RESOURCE_URI}.
    *
    * @see ChannelFinderChannelsIT#handleChannels3QueryByPattern()
    * @see ITTestFixture
